Modified Mutation Delete to be safe

+++ app/graphql/mutations/delete_task.rb
@@ -3,8 +3,16 @@ module Mutations
     field :id, ID, null: true
     argument :id, ID, required: true
     def resolve(id:)
-      task = Task.find(id)
-      task.destroy
+      begin
+        task = Task.find_by(id: id)
+        unless task.nil?
+          task.destroy
+        end
+      rescue ActiveRecord::RecordInvalid => e
+        GraphQL::ExecutionError.new("Invalid attributes for #{e.record.class}:"\
+          " #{e.record.errors.full_messages.join(', ')}")
+      end
+
     end
   end
 end
